% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_df.R
\name{convertIDs}
\alias{convertIDs}
\title{Get other Ids for genes in the same order}
\usage{
convertIDs(ids, from, to, db, ifMultiple = c("putNA", "useFirst"))
}
\arguments{
\item{ids, }{character vector of gene ids}

\item{from, }{type of ID of ids vector}

\item{to, }{type of ID to convert}

\item{db, }{org.species.eg.db class}

\item{ifMultiple, }{what to do when multiple possibles IDs}
}
\value{
annotated dataframe
}
\description{
Get other Ids for genes in the same order
}
\examples{
convertIDs(list_genes, "ENSEMBLTRANS", "SYMBOL", org.Mm.eg.db, "useFirst")
}
\seealso{
\code{\link{AnnotationDbi::select}} used to annotate the dataframe
}

