% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{create_report}
\alias{create_report}
\title{Create report of QC}
\usage{
create_report(path, samples, out, title, condition)
}
\arguments{
\item{path}{absolute path to bcbio results (in yaml file after -upload)}

\item{samples}{vector indicating the name as samples are in the previous folder}

\item{out}{path  where you want to create the index.html}

\item{title}{this value would be the title of the document}

\item{condition}{a vector indicating the groups of the samples}
}
\description{
Create report of QC
}
\examples{
\dontrun{
p<-"absolute paht to bcbio results in yaml file: -upload
s<-c("name1","name2","name3","name4")
o<-"/path/where/html/will/be/created"
t<-"some title for the report"
con<-c("TREAT","TREAT","CON","CON")
r<-create_report(p,s,o,t,con)
}
}

