% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_df.R
\name{get_biomart}
\alias{get_biomart}
\title{Get biomart annotation from id}
\usage{
get_biomart(v, type = "description", id = "ensembl_gene_id",
  dataset = "hsapiens_gene_ensembl")
}
\arguments{
\item{v}{vector of ids}

\item{type}{character name to get from biomart}

\item{id}{character defining what kind of id is provided}

\item{dataset}{specie database to connect to biomart}
}
\value{
annotated vector in the same order
}
\description{
Get biomart annotation from id
}
\examples{
get_biomrt(v, "gene_biotype")
}
\seealso{
\code{\link{biomaRt}} used to annotate the dataframe
}

