% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{mds}
\alias{mds}
\title{Plot MDS}
\usage{
mds(counts, condition = NULL, k = 6, d = "euclidian", xi = 1, yi = 2)
}
\arguments{
\item{counts}{matrix}

\item{condition}{vector indicating the group of each sample}

\item{k}{number of dimension}

\item{d}{distance method (euclidian, cor)}

\item{xi}{PC for x-axis (1 to k)}

\item{yi}{PC for y-axis (1 to k)}
}
\description{
Plot MDS
}

