% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGO.R
\name{print_enrichGO}
\alias{print_enrichGO}
\title{Clean and print results from enrichGO}
\usage{
print_enrichGO(ego, limit = 100)
}
\arguments{
\item{ego}{result object from enrichGO}

\item{limit}{integer limiting the number
of genes in each category. Bigger names
will give broad categories.}
}
\description{
Clean and print results from enrichGO
}

