% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rleoutliers}
\alias{rleoutliers}
\title{Cel file IDs of intensity RLE outlier arrays (in characters).}
\usage{
rleoutliers(eset = NULL, logtransform = TRUE)
}
\arguments{
\item{eset}{AffyBatch}

\item{logtransform}{Boolean, log transform data before processing?}
}
\value{
character vector of arrays (CEL file IDs) that are outliers for intensity boxplots.
}
\description{
Cel file IDs of intensity RLE outlier arrays (in characters).
}
\examples{
rleoutliers(eset=AffyRaw, logtranform=TRUE)
}
\seealso{
\code{\link{arrayQualityMetrics}} which this function subsets
}

