% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGO.R
\name{runGO}
\alias{runGO}
\title{Run GO using clusterprofiler}
\usage{
runGO(genes, org, from = "SYMBOL", ont = "BP", universe = NULL)
}
\arguments{
\item{genes}{character vector with list of genes}

\item{org}{OrgDb object (org.Mm.eg.db)}

\item{from}{character indicanting the name gene type.
Options at: (keytypes(org.Mm.eg.db))}

\item{ont}{ontology to do, BP, MF, CC}

\item{universe}{character vector with list of genes to use as universe}
}
\description{
Run GO using clusterprofiler
}

