% Generated by roxygen2 (4.0.1): do not edit by hand
\name{variance_by_component}
\alias{variance_by_component}
\title{Plot PC importance}
\usage{
variance_by_component(counts, k = 6, d = "euclidian")
}
\arguments{
\item{counts}{matrix}

\item{k}{number of dimension}

\item{d}{distance method (euclidian, cor)}
}
\description{
Plot PC importance
}

