% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTMLreport.R
\name{write_report}
\alias{write_report}
\title{Write report of QC}
\usage{
write_report(report, outfile)
}
\arguments{
\item{report}{an object comming from create_report()}

\item{out}{path where you want to create the index.html (same as used in create_report)}
}
\description{
Write report of QC
}

