% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col2rownames.R
\name{col2rownames}
\alias{col2rownames}
\title{add a column to a dataframe consisting of the current rownames}
\usage{
col2rownames(df, colname, removecol = TRUE)
}
\arguments{
\item{df}{dataframe}

\item{colname}{character vector specifying id of column with new rownames}

\item{removecol}{boolean specifying wheter to remove the column used to rename rownames}
}
\value{
dataframe with column contents as rownames
}
\description{
add a column to a dataframe consisting of the current rownames
}
\examples{
col2rownames(dataframe, "columnid", FALSE)
}

