% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{getTFpairs}
\alias{getTFpairs}
\title{Get mouse TF-target pairs from opossum database}
\usage{
getTFpairs(listgenes, filesqlite, upstream = 10000, downstream = 0)
}
\arguments{
\item{listgene}{vector with ensembl ID genes}

\item{upstream}{limit distance of the TF to the TSS
(upstream to the TSS).}

\item{downstream}{limit distance of the TF to the TSS
(downstream to the TSS).}

\item{filesqlite}{sqlite file containgin the DB.}
}
\value{
data.frame with TF-target pairs
}
\description{
Get mouse TF-target pairs from opossum database
}

