% Generated by roxygen2 (4.0.1): do not edit by hand
\name{write_report}
\alias{write_report}
\title{Write report of QC}
\usage{
write_report(report, outfile)
}
\arguments{
\item{report}{an object comming from create_report()}

\item{out}{path where you want to create the index.html (same as used in create_report)}
}
\description{
Write report of QC
}
\examples{
\dontrun{
p<-"absolute path to bcbio results in yaml file: -upload
s<-c("name1","name2","name3","name4")
o<-"/path/where/html/will/be/created"
t<-"some title for the report"
con<-c("TREAT","TREAT","CON","CON")
r<-write_report(p,s,o,t,con)
}
}

